Split tests from program.
